Add layer-based coloring system
This completely changes the way colors are computed in the coloring
shader, making use of a system of layers to determine both a base color
and lighting, in addition to a couple of overlays. It also adds the
ability to the compute shader to evaluate more sophisticated orbit traps
and stripe average values.

For now, the UI is not hooked up to the new type for coloring settings,
as it is significantly more complex.