Add support for floatexp-like calculation
This adds support for using a separate exponential scaling factor in the
perturbed calculation shader. This fixes two major issues:

- loss of precision when zooming in past 2^-120
- overflow when calculating the derivative for distance estimation

There are still some issues present, most notably a star-shaped glitch
appearing intermittently past 2^-375. It appears to be related to the
probe point's location and the scaling factor used for adjusting the
scaling exponent.